requirements.txt
opencv-python
pillow
tensorflow
keras
moviepy==1.0.3
imageio
tqdm
